feat: callbacks in separate file + results logging in txt #57
